Allow automatic deletion of results files if the test passes.
If the test runs and passes then the default behaviour is to retain
the results files for the last 10 tests. If the first test fails, but
the next 10 pass, then the failed test's results will be deleted.

This functionality allows the user to prevent that by deleting the
results from passing tests.

Signed-off-by: Adam Farley <adam.farley@uk.ibm.com>