Replace SecurityManager use with Java agent
Starting in Java 24 the SecurityManager will be removed from the JDK.
Load tests rely on the SecurityManager to intercept calls to
System.exit and replace them with an exception to avoid
shutting down before all the tests have run. I replaced this with a Java agent
that uses ASM to rewrite System.exit calls to throw an exception instead.
This change also works for previous Java versions.

This change also updates the ASM library version from 9.0 to 9.7.1 to include
Java 24 support. It also adds the asm-commons library as a dependency which is
needed to use the org.objectweb.asm.commons.LocalVariablesSorter library.

Signed-off-by: Theresa Mammarella <Theresa.T.Mammarella@ibm.com>